feat: error handling on failure
